requirements.txt
python-dotenv
requests
paho-mqtt